Create Calculation
There is an option to create or load an old calculation. If the user chooses to create one, they will be sent on the journey of inputting a space decomposition, a velocity set, and start/end points. This will then be put in the optimization method -- the result is the calculation.

The option to load has a construction message standing in its place.


The Elvis program will call on this class to "handle" a selected calculation. We have no such features as of yet, so a construction message is standing in their place.