Update Velocities.py
Now we can actually create real number velocities. When we change to generic convex velocity sets, the entire method will have to be rewritten. Instead of checking if the input is a positive real number, it will check if it is a hyperspherical coordinate function with the right arguments -- and that it is actually convex.

**gulp**